chore: 공용 tailwind config 사용
