+++ packages/tailwind-config/package.json
@@ -4,8 +4,8 @@
   "author": "이호연",
   "license": "MIT",
   "exports": {
-    "./tailwind.config.ts": "./tailwind.config.ts",
-    "./postcss.config.ts": "./postcss.config.ts"
+    "./tailwind.config.js": "./tailwind.config.js",
+    "./postcss.config.js": "./postcss.config.js"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.19",
