+++ packages/tailwind-config/postcss.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: {
+    'tailwindcss/nesting': 'postcss-nesting',
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+};
