--- packages/tailwind-config/postcss.config.ts
@@ -1,9 +1,0 @@
-import config from './tailwind.config';
-
-export default {
-  plugins: {
-    'tailwindcss/nesting': 'postcss-nesting',
-    tailwindcss: { config },
-    autoprefixer: {},
-  },
-};
