+++ packages/ui/postcss.config.cjs
@@ -1,0 +1,1 @@
+module.exports = require('@swifty/tailwind-config/postcss.config.js');
