--- packages/ui/postcss.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  plugins: {
-    'tailwindcss/nesting': 'postcss-nesting',
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
