+++ packages/ui/tailwind.config.cjs
@@ -1,0 +1,1 @@
+module.exports = require('@swifty/tailwind-config/tailwind.config.js');
